\name{cat_function}
\alias{cat_function}
\title{custom title goes here}
\description{for description, function arguments, and return values.
cat_function descr goes here \eqn{a}}
\usage{cat_function(love = TRUE)}\arguments{
  \item{love}{love is a parameter
for documenting function arguments on the same line.
for description, function arguments, and return values.
for documenting function arguments on the same line.}
}

\value{for description, function arguments, and return values.
the return value is a print statement  }

\author{Who wrote it}




\examples{
# comment in example
  cat_function(love = TRUE)
  cat_function(l = FALSE)
}
