\name{catPkg-package}
\alias{catPkg-package}
\alias{catPkg}
\docType{package}
\title{
Sample test package
}
\description{
See http://hilaryparker.com/2014/04/29/writing-an-r-package-from-scratch/
}
\details{
\tabular{ll}{
Package: \tab catPkg\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-12-07\cr
License: \tab What license is it under?\cr
}
Details go here.
}
\author{
Who wrote it - me.

Maintainer: Who to complain to <yourfault@somewhere.net>
}
\references{
http://hilaryparker.com/2014/04/29/writing-an-r-package-from-scratch/
}
\keyword{ package }
%\seealso{
%%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%%~~ simple examples of the most important functions ~~
%}
