\name{catPkg-package}
\alias{catPkg-package}
\alias{catPkg}
\docType{package}
\title{What the package does (short line)}
\description{More about what it does (maybe more than one line)}
\details{
\tabular{ll}{Package: \tab catPkg\cr
Type: \tab Package\cr
Title: \tab What the package does (short line)\cr
Version: \tab 1.0\cr
Date: \tab 2014-12-07\cr
Author: \tab Who wrote it\cr
Maintainer: \tab John Gavin <john.b.gavin+catPkgMaintainer@gmail.com>\cr
Suggests: \tab 
knitr\cr
VignetteBuilder: \tab knitr\cr
License: \tab GPL-3\cr}
}
\author{Who wrote it}

\keyword{ package }


